Factor out the reference file makers
