+++ src/roman_datamodels/maker_utils/__init__.py
@@ -1,3 +1,4 @@
 from ._basic_meta import *  # noqa: F403
 from ._common_meta import *  # noqa: F403
 from ._maker_utils import *  # noqa: F403
+from ._ref_files import *  # noqa: F403
