+++ src/roman_datamodels/maker_utils/_maker_utils.py
@@ -4,7 +4,6 @@ import asdf
 import astropy.time as time
 import numpy as np
 from astropy import units as u
-from astropy.modeling import models
 
 from roman_datamodels import stnode
 from roman_datamodels.random_utils import generate_string
@@ -67,18 +66,6 @@ def mk_resample():
     res["weight_type"] = "exptime"
 
     return res
-
-
-def add_ref_common(meta):
-    instrument = {"name": "WFI", "detector": "WFI01", "optical_element": "F158"}
-    meta["telescope"] = "ROMAN"
-    meta["instrument"] = instrument
-    meta["origin"] = "STSCI"
-    meta["pedigree"] = "GROUND"
-    meta["author"] = "test system"
-    meta["description"] = "blah blah blah"
-    meta["useafter"] = time.Time("2020-01-01T00:00:00.0", format="isot", scale="utc")
-    meta["reftype"] = ""
 
 
 def mk_level1_science_raw(shape=(8, 4096, 4096), filepath=None):
@@ -235,442 +222,6 @@ def mk_level3_mosaic(shape=None, n_images=2, filepath=None):
         af.write_to(filepath)
     else:
         return wfi_mosaic
-
-
-def mk_flat(shape=(4096, 4096), filepath=None):
-    """
-    Create a dummy Flat instance (or file) with arrays and valid values for attributes
-    required by the schema.
-
-    Parameters
-    ----------
-    shape
-        (optional) Shape of arrays in the model.
-
-    filepath
-        (optional) File name and path to write model to.
-
-    Returns
-    -------
-    roman_datamodels.stnode.FlatRef
-    """
-    meta = {}
-    add_ref_common(meta)
-    flatref = stnode.FlatRef()
-    meta["reftype"] = "FLAT"
-    flatref["meta"] = meta
-
-    flatref["data"] = np.zeros(shape, dtype=np.float32)
-    flatref["dq"] = np.zeros(shape, dtype=np.uint32)
-    flatref["err"] = np.zeros(shape, dtype=np.float32)
-
-    if filepath:
-        af = asdf.AsdfFile()
-        af.tree = {"roman": flatref}
-        af.write_to(filepath)
-    else:
-        return flatref
-
-
-def mk_dark(shape=(2, 4096, 4096), filepath=None):
-    """
-    Create a dummy Dark Current instance (or file) with arrays and valid values for attributes
-    required by the schema.
-
-    Parameters
-    ----------
-    shape
-        (optional) Shape of arrays in the model.
-
-    filepath
-        (optional) File name and path to write model to.
-
-    Returns
-    -------
-    roman_datamodels.stnode.DarkRef
-    """
-    meta = {}
-    add_ref_common(meta)
-    darkref = stnode.DarkRef()
-    meta["reftype"] = "DARK"
-    darkref["meta"] = meta
-    exposure = {}
-    exposure["ngroups"] = 6
-    exposure["nframes"] = 8
-    exposure["groupgap"] = 0
-    exposure["type"] = "WFI_IMAGE"
-    exposure["p_exptype"] = "WFI_IMAGE|WFI_GRISM|WFI_PRISM|"
-    exposure["ma_table_name"] = NOSTR
-    exposure["ma_table_number"] = NONUM
-    darkref["meta"]["exposure"] = exposure
-
-    darkref["data"] = u.Quantity(np.zeros(shape, dtype=np.float32), u.DN, dtype=np.float32)
-    darkref["dq"] = np.zeros(shape[1:], dtype=np.uint32)
-    darkref["err"] = u.Quantity(np.zeros(shape, dtype=np.float32), u.DN, dtype=np.float32)
-
-    if filepath:
-        af = asdf.AsdfFile()
-        af.tree = {"roman": darkref}
-        af.write_to(filepath)
-    else:
-        return darkref
-
-
-def mk_distortion(filepath=None):
-    """
-    Create a dummy Distortion instance (or file) with arrays and valid values for attributes
-    required by the schema.
-
-    Parameters
-    ----------
-
-    filepath
-        (optional) File name and path to write model to.
-
-    Returns
-    -------
-    roman_datamodels.stnode.DistortionRef
-    """
-    meta = {}
-    add_ref_common(meta)
-    distortionref = stnode.DistortionRef()
-    meta["reftype"] = "DISTORTION"
-    distortionref["meta"] = meta
-
-    distortionref["meta"]["input_units"] = u.pixel
-    distortionref["meta"]["output_units"] = u.arcsec
-
-    distortionref["coordinate_distortion_transform"] = models.Shift(1) & models.Shift(2)
-
-    if filepath:
-        af = asdf.AsdfFile()
-        af.tree = {"roman": distortionref}
-        af.write_to(filepath)
-    else:
-        return distortionref
-
-
-def mk_gain(shape=(4096, 4096), filepath=None):
-    """
-    Create a dummy Gain instance (or file) with arrays and valid values for attributes
-    required by the schema.
-
-    Parameters
-    ----------
-    shape
-        (optional) Shape of arrays in the model.
-
-    filepath
-        (optional) File name and path to write model to.
-
-    Returns
-    -------
-    roman_datamodels.stnode.GainRef
-    """
-    meta = {}
-    add_ref_common(meta)
-    gainref = stnode.GainRef()
-    meta["reftype"] = "GAIN"
-    gainref["meta"] = meta
-
-    gainref["data"] = u.Quantity(np.zeros(shape, dtype=np.float32), u.electron / u.DN, dtype=np.float32)
-
-    if filepath:
-        af = asdf.AsdfFile()
-        af.tree = {"roman": gainref}
-        af.write_to(filepath)
-    else:
-        return gainref
-
-
-def mk_ipc(shape=(3, 3), filepath=None):
-    """
-    Create a dummy IPC instance (or file) with arrays and valid values for attributes
-    required by the schema.
-
-    Parameters
-    ----------
-    shape
-        (optional) Shape of array in the model.
-
-    filepath
-        (optional) File name and path to write model to.
-
-    Returns
-    -------
-    roman_datamodels.stnode.IpcRef
-    """
-    meta = {}
-    add_ref_common(meta)
-    ipcref = stnode.IpcRef()
-    meta["reftype"] = "IPC"
-    ipcref["meta"] = meta
-
-    ipcref["data"] = np.zeros(shape, dtype=np.float32)
-    ipcref["data"][int(np.floor(shape[0] / 2))][int(np.floor(shape[1] / 2))] = 1.0
-
-    if filepath:
-        af = asdf.AsdfFile()
-        af.tree = {"roman": ipcref}
-        af.write_to(filepath)
-    else:
-        return ipcref
-
-
-def mk_linearity(shape=(2, 4096, 4096), filepath=None):
-    """
-    Create a dummy Linearity instance (or file) with arrays and valid values for attributes
-    required by the schema.
-
-    Parameters
-    ----------
-    shape
-        (optional) Shape of arrays in the model.
-
-    filepath
-        (optional) File name and path to write model to.
-
-    Returns
-    -------
-    roman_datamodels.stnode.LinearityRef
-    """
-    meta = {}
-    add_ref_common(meta)
-    linearityref = stnode.LinearityRef()
-    meta["reftype"] = "LINEARITY"
-    linearityref["meta"] = meta
-
-    linearityref["meta"]["input_units"] = u.DN
-    linearityref["meta"]["output_units"] = u.DN
-
-    linearityref["dq"] = np.zeros(shape[1:], dtype=np.uint32)
-    linearityref["coeffs"] = np.zeros(shape, dtype=np.float32)
-
-    if filepath:
-        af = asdf.AsdfFile()
-        af.tree = {"roman": linearityref}
-        af.write_to(filepath)
-    else:
-        return linearityref
-
-
-def mk_inverse_linearity(shape=(2, 4096, 4096), filepath=None):
-    """
-    Create a dummy InverseLinearity instance (or file) with arrays and valid
-    values for attributes required by the schema.
-
-    Parameters
-    ----------
-    shape
-        (optional) Shape of arrays in the model.
-
-    filepath
-        (optional) File name and path to write model to.
-
-    Returns
-    -------
-    roman_datamodels.stnode.InverseLinearityRef
-    """
-    meta = {}
-    add_ref_common(meta)
-    inverselinearityref = stnode.InverseLinearityRef()
-    meta["reftype"] = "INVERSELINEARITY"
-    inverselinearityref["meta"] = meta
-
-    inverselinearityref["meta"]["input_units"] = u.DN
-    inverselinearityref["meta"]["output_units"] = u.DN
-
-    inverselinearityref["dq"] = np.zeros(shape[1:], dtype=np.uint32)
-    inverselinearityref["coeffs"] = np.zeros(shape, dtype=np.float32)
-
-    if filepath:
-        af = asdf.AsdfFile()
-        af.tree = {"roman": inverselinearityref}
-        af.write_to(filepath)
-    else:
-        return inverselinearityref
-
-
-def mk_mask(shape=(4096, 4096), filepath=None):
-    """
-    Create a dummy Mask instance (or file) with arrays and valid values for attributes
-    required by the schema.
-
-    Parameters
-    ----------
-    shape
-        (optional) Shape of arrays in the model.
-
-    filepath
-        (optional) File name and path to write model to.
-
-    Returns
-    -------
-    roman_datamodels.stnode.MaskRef
-    """
-    meta = {}
-    add_ref_common(meta)
-    maskref = stnode.MaskRef()
-    meta["reftype"] = "MASK"
-    maskref["meta"] = meta
-
-    maskref["dq"] = np.zeros(shape, dtype=np.uint32)
-
-    if filepath:
-        af = asdf.AsdfFile()
-        af.tree = {"roman": maskref}
-        af.write_to(filepath)
-    else:
-        return maskref
-
-
-def mk_pixelarea(shape=(4096, 4096), filepath=None):
-    """
-    Create a dummy Pixelarea instance (or file) with arrays and valid values for attributes
-    required by the schema.
-
-    Parameters
-    ----------
-    shape
-        (optional) Shape of arrays in the model.
-
-    filepath
-        (optional) File name and path to write model to.
-
-    Returns
-    -------
-    roman_datamodels.stnode.PixelareaRef
-    """
-    meta = {}
-    add_ref_common(meta)
-    pixelarearef = stnode.PixelareaRef()
-    meta["reftype"] = "AREA"
-    meta["photometry"] = {
-        "pixelarea_steradians": float(NONUM) * u.sr,
-        "pixelarea_arcsecsq": float(NONUM) * u.arcsec**2,
-    }
-    pixelarearef["meta"] = meta
-
-    pixelarearef["data"] = np.zeros(shape, dtype=np.float32)
-
-    if filepath:
-        af = asdf.AsdfFile()
-        af.tree = {"roman": pixelarearef}
-        af.write_to(filepath)
-    else:
-        return pixelarearef
-
-
-def mk_wfi_img_photom(filepath=None):
-    """
-    Create a dummy WFI Img Photom instance (or file) with dictionary and valid values for attributes
-    required by the schema.
-
-    Parameters
-    ----------
-    filepath
-        (optional) File name and path to write model to.
-
-    Returns
-    -------
-    roman_datamodels.stnode.WfiImgPhotomRef
-    """
-    meta = {}
-    add_ref_common(meta)
-    wfi_img_photomref = stnode.WfiImgPhotomRef()
-    meta["reftype"] = "PHOTOM"
-    wfi_img_photomref["meta"] = meta
-
-    wfi_img_photo_dict = {
-        "F062": {
-            "photmjsr": (1.0e-15 * np.random.random() * u.megajansky / u.steradian),
-            "uncertainty": (1.0e-16 * np.random.random() * u.megajansky / u.steradian),
-            "pixelareasr": 1.0e-13 * u.steradian,
-        },
-        "F087": {
-            "photmjsr": (1.0e-15 * np.random.random() * u.megajansky / u.steradian),
-            "uncertainty": (1.0e-16 * np.random.random() * u.megajansky / u.steradian),
-            "pixelareasr": 1.0e-13 * u.steradian,
-        },
-        "F106": {
-            "photmjsr": (1.0e-15 * np.random.random() * u.megajansky / u.steradian),
-            "uncertainty": (1.0e-16 * np.random.random() * u.megajansky / u.steradian),
-            "pixelareasr": 1.0e-13 * u.steradian,
-        },
-        "F129": {
-            "photmjsr": (1.0e-15 * np.random.random() * u.megajansky / u.steradian),
-            "uncertainty": (1.0e-16 * np.random.random() * u.megajansky / u.steradian),
-            "pixelareasr": 1.0e-13 * u.steradian,
-        },
-        "F146": {
-            "photmjsr": (1.0e-15 * np.random.random() * u.megajansky / u.steradian),
-            "uncertainty": (1.0e-16 * np.random.random() * u.megajansky / u.steradian),
-            "pixelareasr": 1.0e-13 * u.steradian,
-        },
-        "F158": {
-            "photmjsr": (1.0e-15 * np.random.random() * u.megajansky / u.steradian),
-            "uncertainty": (1.0e-16 * np.random.random() * u.megajansky / u.steradian),
-            "pixelareasr": 1.0e-13 * u.steradian,
-        },
-        "F184": {
-            "photmjsr": (1.0e-15 * np.random.random() * u.megajansky / u.steradian),
-            "uncertainty": (1.0e-16 * np.random.random() * u.megajansky / u.steradian),
-            "pixelareasr": 1.0e-13 * u.steradian,
-        },
-        "F213": {
-            "photmjsr": (1.0e-15 * np.random.random() * u.megajansky / u.steradian),
-            "uncertainty": (1.0e-16 * np.random.random() * u.megajansky / u.steradian),
-            "pixelareasr": 1.0e-13 * u.steradian,
-        },
-        "GRISM": {"photmjsr": None, "uncertainty": None, "pixelareasr": 1.0e-13 * u.steradian},
-        "PRISM": {"photmjsr": None, "uncertainty": None, "pixelareasr": 1.0e-13 * u.steradian},
-        "DARK": {"photmjsr": None, "uncertainty": None, "pixelareasr": 1.0e-13 * u.steradian},
-    }
-    wfi_img_photomref["phot_table"] = wfi_img_photo_dict
-
-    if filepath:
-        af = asdf.AsdfFile()
-        af.tree = {"roman": wfi_img_photomref}
-        af.write_to(filepath)
-    else:
-        return wfi_img_photomref
-
-
-def mk_readnoise(shape=(4096, 4096), filepath=None):
-    """
-    Create a dummy Readnoise instance (or file) with arrays and valid values for attributes
-    required by the schema.
-
-    Parameters
-    ----------
-    shape
-        (optional) Shape of arrays in the model.
-
-    filepath
-        (optional) File name and path to write model to.
-
-    Returns
-    -------
-    roman_datamodels.stnode.ReadnoiseRef
-    """
-    meta = {}
-    add_ref_common(meta)
-    readnoiseref = stnode.ReadnoiseRef()
-    meta["reftype"] = "READNOISE"
-    readnoiseref["meta"] = meta
-    exposure = {}
-    exposure["type"] = "WFI_IMAGE"
-    exposure["p_exptype"] = "WFI_IMAGE|WFI_GRISM|WFI_PRISM|"
-    readnoiseref["meta"]["exposure"] = exposure
-
-    readnoiseref["data"] = u.Quantity(np.zeros(shape, dtype=np.float32), u.DN, dtype=np.float32)
-
-    if filepath:
-        af = asdf.AsdfFile()
-        af.tree = {"roman": readnoiseref}
-        af.write_to(filepath)
-    else:
-        return readnoiseref
 
 
 def mk_ramp(shape=(8, 4096, 4096), filepath=None):
@@ -872,72 +423,3 @@ def mk_guidewindow(shape=(2, 8, 16, 32, 32), filepath=None):
         af.write_to(filepath)
     else:
         return guidewindow
-
-
-def mk_saturation(shape=(4096, 4096), filepath=None):
-    """
-    Create a dummy Saturation instance (or file) with arrays and valid values for attributes
-    required by the schema.
-
-    Parameters
-    ----------
-    shape
-        (optional) Shape of arrays in the model.
-
-    filepath
-        (optional) File name and path to write model to.
-
-    Returns
-    -------
-    roman_datamodels.stnode.SaturationRef
-    """
-    meta = {}
-    add_ref_common(meta)
-    saturationref = stnode.SaturationRef()
-    meta["reftype"] = "SATURATION"
-    saturationref["meta"] = meta
-
-    saturationref["dq"] = np.zeros(shape, dtype=np.uint32)
-    saturationref["data"] = u.Quantity(np.zeros(shape, dtype=np.float32), u.DN, dtype=np.float32)
-
-    if filepath:
-        af = asdf.AsdfFile()
-        af.tree = {"roman": saturationref}
-        af.write_to(filepath)
-    else:
-        return saturationref
-
-
-def mk_superbias(shape=(4096, 4096), filepath=None):
-    """
-    Create a dummy Superbias instance (or file) with arrays and valid values for attributes
-    required by the schema.
-
-    Parameters
-    ----------
-    shape
-        (optional) Shape of arrays in the model.
-
-    filepath
-        (optional) File name and path to write model to.
-
-    Returns
-    -------
-    roman_datamodels.stnode.SuperbiasRef
-    """
-    meta = {}
-    add_ref_common(meta)
-    superbiasref = stnode.SuperbiasRef()
-    meta["reftype"] = "BIAS"
-    superbiasref["meta"] = meta
-
-    superbiasref["data"] = np.zeros(shape, dtype=np.float32)
-    superbiasref["dq"] = np.zeros(shape, dtype=np.uint32)
-    superbiasref["err"] = np.zeros(shape, dtype=np.float32)
-
-    if filepath:
-        af = asdf.AsdfFile()
-        af.tree = {"roman": superbiasref}
-        af.write_to(filepath)
-    else:
-        return superbiasref
